+++ addon/array/index.js
@@ -20,3 +20,6 @@ export { default as map } from './map';
 export { default as objectAt } from './object-at';
 export { default as reduce } from './reduce';
 export { default as slice } from './slice';
+export { default as uniqBy } from './uniq-by';
+export { default as uniq } from './uniq';
+export { default as without } from './without';
