+++ addon/array/without.js
@@ -1,4 +1,4 @@
-import { normalizeArray } from './utils';
+import { normalizeArray } from '../utils';
 
 export default function(...keys) {
   return normalizeArray(keys, {}, (array, item) => {
